[GEN-2042]: do not display warning when closing drawer of "unknown instrumentation rule" (#2036)
This pull request includes a small but important change to the
`RuleDrawer` component in the
`frontend/webapp/containers/main/instrumentation-rules/rule-drawer/index.tsx`
file. The change adds an additional condition to the `handleEdit`
function to improve the handling of unknown rule types.

*
[`frontend/webapp/containers/main/instrumentation-rules/rule-drawer/index.tsx`](diffhunk://#diff-759940924dd4ee0c6895f60e21ca040542d9cf1d3f69162677c9983d62b12f8cL76-R76):
Modified the `handleEdit` function to include a condition that checks if
the `bool` parameter is `true` or `undefined` before adding a
notification for unknown rule types.