refactor: use instrumentationconfig in autoscaler (#2025)
Part of the effort to retire instrumentedapplication

This object behaves the same, so no changes in functionality is
expected.
Updated the RBAC roles and cleaned up unused permissions while at it (we
need to do a bigger rbac swap in a different PR).

Tested it to work with both cli and helm (with logs destination that
triggers change in collector cm to collect logs).