chore: narrow the RBAC permissions used by the autoscaler (#2039)
This reduces the permissions we request to just those we actually use.

Most resources are moved from cluster role to role. only resource we
need in cluster level is `instrumentationconfigs` to update the
datacollection configmaps with file patterns to scrape for logs.

Updated the selectors in controller runtime cache.

Tested several flows with CLI and helm.