[GEN-2096]: conditionally display runtime version in source containers (#2057)
This pull request includes a change to the
`frontend/webapp/reuseable-components/data-card/data-card-fields/index.tsx`
file. The change improves the subtitle rendering logic for the `DataTab`
component by conditionally appending the runtime version only if it
exists.

*
[`frontend/webapp/reuseable-components/data-card/data-card-fields/index.tsx`](diffhunk://#diff-ce146776b402d3e7c41f2ebb82b32a2b8636ec73200d70a6f30f0f08dc1da6d0L99-R99):
Modified the `renderValue` function to append the runtime version to the
subtitle only if it is defined.

---------

Signed-off-by: Ben Elferink <ben.elferink@icloud.com>