Introduce Source CRD (#2037)
This adds the `Source` CRD that will be used to enable auto
instrumentation for a workload (vs the current label approach). Based on
[design
doc](https://www.notion.so/keyval/Instrumentation-CRD-design-15e5ac7207b180bc9c94fd441af94176?pvs=4)

This doesn't hook the CRD up to anything yet or enable any functionality
based on it. That will be done in a feature branch to collaborate with
frontend. This PR is meant to break the change into a logical chunk, but
also to unblock other work that can rely on this API type.