[GEN-2110]: update RBAC permissions for UI (#2058)
This pull request includes significant updates to the role and cluster
role definitions in the `cli/cmd/resources/ui.go` file. The changes
primarily involve refining the permissions for various resources to
better align with the required operations.

### Changes to Role and ClusterRole Definitions:

*
[`cli/cmd/resources/ui.go`](diffhunk://#diff-c286e10d34710a80a59127b2b7951e8a33d9b9554e47d2f2b827fd690f2e53abL143-R156):
Updated the `NewUIRole` function to refine the permissions for accessing
`configmaps`, `secrets`, and `odigos.io` resources. This includes adding
comments to clarify the necessity of each permission set.
*
[`cli/cmd/resources/ui.go`](diffhunk://#diff-c286e10d34710a80a59127b2b7951e8a33d9b9554e47d2f2b827fd690f2e53abL231-R230):
Updated the `NewUIClusterRole` function to refine the permissions for
accessing `namespaces`, `services`, `pods`, `replicasets`, and
`odigos.io` resources. Comments were added to explain the purpose of
each permission set.