Implement Source CRD instrumentation (#2059)
This introduces the basic implementation for
instrumenting/uninstrumenting via `Source` objects
(https://github.com/odigos-io/odigos/pull/2037)

This does not remove current instrumentation with the
`odigos-instrumentation` label, and does not do any checks to avoid
conflicts etc. That work will follow. Right now this just drops in the
basic workflow alongside the current label approach.

Changes:

* Add a `SourceReconciler` to the `startlangdetection` Instrumentor
controller.
  * Parse workload from `Source` object
* If the `Source` is not being deleted, add 2 finalizers (one used for
`Source` deletion, and another used by the
`deleteinstrumentedapplication` controller), add workload labels to the
`Source` for lookup, and continue with flow of creating an
`InstrumentationConfig` by calling
`requestOdigletsToCalculateRuntimeDetails`
* If the `Source` is being deleted (determined by a `DeletionTimestamp`
being present in an Update event, with the finalizer blocking deletion),
remove the source finalizer and delete the `InstrumentationConfig`.
* Add a function `GetSourceListForWorkload` everywhere we currently call
`workload.IsWorkloadInstrumentationEffectiveEnabled` that returns any
`Source` objects referencing the workload.
* Add `SourceReconciler` to `deleteinstrumentedapplication` Instrumentor
controller
* When `Source` object is deleted, remove instrumented application
finalizer and continue with reconciling the workload, which should
trigger uninstrumentation based on the new checks for
`GetSourceListForWorkload` throughout

I'm not a fan of using 2 finalizers and worry that we could end up with
a deadlock/race. This is mostly due to InstrumentedApplication, and
deprecating InstrumentedApplication should simplify the work required
for this greatly.

Still needs:

* Namespace instrumentation and exclusion
* Switch `odigos-instrumentation` label to be shorthand for creating a
`Source` on that workload/namespace
* This includes removing current label checks with
`workload.IsWorkloadInstrumentationEffectiveEnabled`
* UI integration
* Webhook for validation (immutability) and defaulting (labels)