[GEN-2133]: fix height of source-list (#2090)
This pull request includes several changes to improve the handling of
modal and non-modal states in the `ChooseSourcesBody` components and
their styling. The most important changes include modifying the
`ModalBody` component to accept a new prop, updating the
`ChooseSourcesBodyFast` and `ChooseSourcesBodySimple` components to use
this new prop, and adjusting the styling for the `SourcesList`
component.

Changes to `ModalBody` component:

*
[`frontend/webapp/styles/styled.tsx`](diffhunk://#diff-f49af8b262d0ee0275a549c3b301b81f5bacbe5a9287a247832e33851168f99eL24-R27):
Modified the `ModalBody` component to accept a new `$isNotModal` prop
and adjusted its height and margin based on this prop.

Updates to `ChooseSourcesBody` components:

*
[`frontend/webapp/containers/main/sources/choose-sources/choose-sources-body/choose-sources-body-fast/index.tsx`](diffhunk://#diff-7525bbd72ee82a35a9ccee85b94d510bcc84b729ad765e378eee4e3df628a972L13-R13):
Updated the `ChooseSourcesBodyFast` component to pass the new
`$isNotModal` prop to `ModalBody`.
*
[`frontend/webapp/containers/main/sources/choose-sources/choose-sources-body/choose-sources-body-simple/index.tsx`](diffhunk://#diff-c14b373dc9584d63b54caf4c08380c97621527bb4bc20a66d7e73cbef4bbbda7L13-R13):
Updated the `ChooseSourcesBodySimple` component to pass the new
`$isNotModal` prop to `ModalBody`.

Styling adjustments:

*
[`frontend/webapp/containers/main/sources/choose-sources/choose-sources-body/choose-sources-body-fast/sources-list/index.tsx`](diffhunk://#diff-6b67649d370d208941f4e5a78c0c9de2f5b9b65fd49e0ede8ef57066982450aaL17-L19):
Adjusted the `max-height` and removed `padding-bottom` styles based on
the `$isModal` prop.