chore: remove dead test (#2137)
This pull request includes updates to the Go version specification in
the GitHub Actions workflows and the removal of a test suite for the
`instrumentationdevice` controller.

Changes to GitHub Actions workflows:

*
[`.github/workflows/build.yaml`](diffhunk://#diff-d0777657fa3fd81d23aaf7273e58aee453b04e67882517900c56daeef9b3e4c1L46-R46):
Updated the `go-version` specification from double quotes to single
quotes in multiple job steps.
[[1]](diffhunk://#diff-d0777657fa3fd81d23aaf7273e58aee453b04e67882517900c56daeef9b3e4c1L46-R46)
[[2]](diffhunk://#diff-d0777657fa3fd81d23aaf7273e58aee453b04e67882517900c56daeef9b3e4c1L121-R121)
[[3]](diffhunk://#diff-d0777657fa3fd81d23aaf7273e58aee453b04e67882517900c56daeef9b3e4c1L134-R134)
[[4]](diffhunk://#diff-d0777657fa3fd81d23aaf7273e58aee453b04e67882517900c56daeef9b3e4c1L146-R146)
[[5]](diffhunk://#diff-d0777657fa3fd81d23aaf7273e58aee453b04e67882517900c56daeef9b3e4c1L158-R158)

Removal of test suite:

*
[`instrumentor/controllers/instrumentationdevice/suite_test.go`](diffhunk://#diff-ba43311eab39abad82636769e90ee6086c1b750ee814ee92f9e66e5a7b2bc03dL1-L131):
Removed the entire test suite for the `instrumentationdevice`
controller, including setup and teardown logic.