feat: record installation method into odigos-deployment (#2184)
This PR adds "instllation method" to `odigos-deployment` config map, so
we can reliably know if the current installation is deployed via "helm"
or "odigos-cli".

We can later use this info to warn or block usages which we do not
support and might end up with a broken deployment or odigos.

This PR also adds the installation method and tier to the describe cli
command:

```
go run -tags=embed_manifests . describe                           
Odigos Version: v1.0.142
Tier: community
Installation Method: helm

....
```