chore: align our k8s go dep to same version (#2189)
We have a mix of dependencies on `k8s.io/api` modules.

Some modules requires `v0.31.0` and some `v0.32.0`. This caused issues
in #2187 , so want to clean it up in a seprate PR and then rebase.

We need to understand why dependabot updated just some of them and not
all