chore: create "profiles" module and migrate part of code to there (#2187)
Current situation:

- Our available profiles are set in `k8sutils`
- The yamls they need to apply is in `cli`
- supported profiles per tier is in `cli`
- profiles are used from: cli, schedualer, ui

This PR creates a new `profiles` go module, similar to the one we
currently have for `destinations` and migrate part of the code to there.
It removes everything profile related from `k8sutils` which are not a
good place to store the actual profiles and maintain them in utils
package that is consumed by all other modules.