chore: move profiles manifest files to profiles module (#2193)
profiles manifests are currently hosted in cli module.

this PR moves them to the `profiles` module, so they can later be
consumed and apply as a reconciliation in sschedualer.

It will also help maintainability to have all profiles definitions in
one place instead of having the profile info in one module and manifests
in another