chore: reconcile profiles in schedualer (#2210)
Instead of only calculating them in cli.

This PR makes it so that profiles can be added by helm, ui, gitOps.
anything that will end up updating the profiles list in `odigos-config`
will cause profiles to update in the cluster and get applied (unlike
today where only cli is supported)