[GEN-2231]: apply logic with DisableInstrumentation for namespace and workload sources instrumentation via UI (#2242)
This pull request includes several changes to improve the handling of
Kubernetes namespaces and workloads, focusing on instrumentation status
and the creation and updating of Custom Resource Definitions (CRDs). The
most important changes include updates to the `schema.resolvers.go` file
for namespace and workload instrumentation checks, modifications to the
`sources.go` file for CRD management, and improvements to the
`useSourceCRUD` hook for better user notifications.

Changes to namespace and workload instrumentation checks:

*
[`frontend/graph/schema.resolvers.go`](diffhunk://#diff-8e6e95029056db2c0301fc338e0ca5a04356ce5d45ee9514bbd167f2d85bae70R101-R112):
Updated the `K8sActualNamespace` resolver to check if the entire
namespace is instrumented by fetching the relevant CRD and determining
the instrumentation status.
*
[`frontend/graph/schema.resolvers.go`](diffhunk://#diff-8e6e95029056db2c0301fc338e0ca5a04356ce5d45ee9514bbd167f2d85bae70L320-R346):
Modified the `K8sActualSources` resolver to handle namespace and
workload instrumentation checks more efficiently by refactoring the
logic and improving error handling.

CRD management improvements:

*
[`frontend/services/sources.go`](diffhunk://#diff-d8f62b6675961fc4e307e4fd622a59132ddc730b6e701ae1bc43dd1695f969a7L292-R307):
Updated the `createSourceCRD` function to return the created or updated
source and added a new `updateSourceCRD` function to handle CRD updates.
[[1]](diffhunk://#diff-d8f62b6675961fc4e307e4fd622a59132ddc730b6e701ae1bc43dd1695f969a7L292-R307)
[[2]](diffhunk://#diff-d8f62b6675961fc4e307e4fd622a59132ddc730b6e701ae1bc43dd1695f969a7L321-R331)
*
[`frontend/services/sources.go`](diffhunk://#diff-d8f62b6675961fc4e307e4fd622a59132ddc730b6e701ae1bc43dd1695f969a7R340-R374):
Enhanced the `deleteSourceCRD` function to check for namespace sources
and update workload sources with `DisableInstrumentation` if necessary.

User notification enhancements:

*
[`frontend/webapp/hooks/sources/useSourceCRUD.ts`](diffhunk://#diff-1be86445dc0ed6bfdc08f7525c5800ef39bdcd70318a3c130f83f36f1ae9f5c8L101-R109):
Improved user notifications in the `persistSources` method to provide
more specific feedback when persisting sources and namespaces.
[[1]](diffhunk://#diff-1be86445dc0ed6bfdc08f7525c5800ef39bdcd70318a3c130f83f36f1ae9f5c8L101-R109)
[[2]](diffhunk://#diff-1be86445dc0ed6bfdc08f7525c5800ef39bdcd70318a3c130f83f36f1ae9f5c8L115-R135)