[GEN-2132]: remove old `uninstall` from docs (#2259)
This pull request includes several changes to the documentation files to
improve clarity and organization. The most important changes include
updating the `docs/mint.json` file and modifying the content in
`docs/telemetry-types.mdx` and `docs/uninstall.mdx`.

Changes to documentation structure:

*
[`docs/mint.json`](diffhunk://#diff-c91a604899dfef4b2494c317f4fd39a7f22b79986095f580399347293d534debL92-R92):
Consolidated the `pages` arrays for the "API Reference" group into a
single line and removed the "uninstall" page from the "Concepts" group.
[[1]](diffhunk://#diff-c91a604899dfef4b2494c317f4fd39a7f22b79986095f580399347293d534debL92-R92)
[[2]](diffhunk://#diff-c91a604899dfef4b2494c317f4fd39a7f22b79986095f580399347293d534debL235-R232)

Content updates:

*
[`docs/telemetry-types.mdx`](diffhunk://#diff-4467258929f82d0967c563a3414c162f923821e1f3019300b0083c44ba766884L7-R11):
Simplified the terminology for telemetry types and added a link to the
"destination" page for better navigation.
*
[`docs/uninstall.mdx`](diffhunk://#diff-a05eabe73b3aa6a33cb03155d0630e9cb7b683301204bec9cd6a7d92d52e700dL1-L22):
Removed the entire uninstallation guide and note about the
work-in-progress status of the uninstallation process.