[GEN-2233]: enhance AddEntity component with dropdown hover functionality and style updates (#2261)
This pull request includes several changes to the
`frontend/webapp/components/overview/add-entity/index.tsx` file to
improve the user interface and functionality of the Add Entity
component. The changes include updates to the styles, addition of new
hooks, and modifications to the component's behavior.

### Style Updates:
* Increased the minimum width of the `StyledButton` to 160px and added
padding to the right.
* Changed the `DropdownListContainer`'s top position to 42px and updated
the background color to use `theme.colors.secondary`.
* Modified the hover background color for `DropdownItem` to use
`theme.colors.majestic_blue` with a 50% opacity.

### Functional Updates:
* Added `useEffect` to handle opening and closing the dropdown on mouse
enter and leave events.
* Updated the `AddEntity` component to use the `useTheme` hook and
removed the `handleToggle` function.

### Minor Changes:
* Changed the placeholder text from 'ADD...' to 'ADD NEW'.
* Updated the `PlusIcon` and `Text` components to use the theme's
primary color for fill and text color.