Migrate `odigos-instrumentation` label to Source CRD (#2231)
This adds a migration for the current `odigos-instrumentation` label to
Source CRDs.

If a workload or namespace has `odigos-instrumentation: enabled`, the
startlangdetection controller will automatically create a Source for it
(if one doesn't already exist) and log a deprecation warning.

If a workload or namespace has `odigos-instrumentation: disabled`, the
deleteinstrumentationconfig controller will automatically create a
Source with `disableInstrumentation: true`.

In both cases, if a Source already exists, the controller will update
the existing Source. This means that the label can be toggled back and
forth.

However, deleting the label will not inherently un-instrument a workload
that doesn't have inheriting Namespace instrumentation. We can't tell if
the label was deleted without checking the change during the event,
which is bad to do. This is the only drop in feature parity from the
current label.