chore: move k8s consts from k8sutils module to api module (#2281)
There is currently a dependency of `api` module on `k8sutils` and also
`k8sutils` depends on `api`.

This sometimes creates cyclic import and is a bad practice.
Current PR moves the consts from `k8sutils` to `api` module, as the
first step towards breaking this cycle.

The end result will be having the `k8sutils` depend on `api`, but not
the other way around.

This PR simply moves things between modules, no renames, no changes
(even though there are some things we can address to improve). wanted to
not introduce any changes to this PR that touch so many files all over