[GEN-2239]: sources header node - display "filtered / total" counts (#2299)
This pull request includes changes to improve the display of source
nodes and to correct import statements in the `header-node` component.
The most important changes are as follows:

Improvements to source node display:

*
[`frontend/webapp/containers/main/overview/overview-data-flow/build-source-nodes.ts`](diffhunk://#diff-278da68d858c88d202c10e2011bc1263fe6c0b543dc9d0fc4995a0f0675d3db1L53-R53):
Modified the `tagValue` to display a formatted string showing the number
of filtered entities when it differs from the unfiltered count.

Corrections to import statements and type definitions:

*
[`frontend/webapp/reuseable-components/nodes-data-flow/nodes/header-node.tsx`](diffhunk://#diff-ca44442b51175efdcec98d2bfebf14b5214feb64fdf6c62dec29e5fe20e98ba9L3-R7):
Corrected the order of import statements to follow best practices.
*
[`frontend/webapp/reuseable-components/nodes-data-flow/nodes/header-node.tsx`](diffhunk://#diff-ca44442b51175efdcec98d2bfebf14b5214feb64fdf6c62dec29e5fe20e98ba9L16-R16):
Updated the `tagValue` type in the `Props` interface to accept both
string and number types.