[GEN-2251]: add tooltip to metrics (aka labeled edges) (#2300)
This pull request includes a small enhancement to the `LabeledEdge`
component in the `nodes-data-flow` directory. The change introduces a
tooltip to provide additional context about the displayed data-flow
metrics.

Enhancements to `LabeledEdge` component:

*
[`frontend/webapp/reuseable-components/nodes-data-flow/edges/labeled-edge.tsx`](diffhunk://#diff-6628e1ca597f995b5a57a34621867f9a8651b7b0af3e2f835c6113079d717084R4):
Imported the `Tooltip` component and wrapped the `Label` component with
it to display a message explaining that the data-flow metrics are
calculated over 10-second intervals.
[[1]](diffhunk://#diff-6628e1ca597f995b5a57a34621867f9a8651b7b0af3e2f835c6113079d717084R4)
[[2]](diffhunk://#diff-6628e1ca597f995b5a57a34621867f9a8651b7b0af3e2f835c6113079d717084R51-R55)