Generate CLI docs (#2319)
This adds a root `make cli-docs` command to generate CLI docs using the
cobra provided markdown generator (`scripts/cli-docgen/main.go`).

This also updates the help text and examples of our CLI commands to try
to match what is currently published, now that this text will come
straight from the command itself.

This also adds a github workflow to check if CLI doc updates are needed.
This should all make it much easier to keep our cli docs up to date and
in sync with the commands themselves

---------

Co-authored-by: Ben Elferink <ben.elferink@icloud.com>
Co-authored-by: Amir Blum <amirgiraffe@gmail.com>