Add reported name field to Source, deprecate the odigos.io/reported-name annotation (#2298)
This PR deprecates the `odigos.io/reported-name` annotation, as part of
the effort the reduce the changes we do to workload objects.
* Adding an optional `reportedName` field to the Source spec. This field
is only valid for workloads Sources and not for Namesoace Sources. This
is enforced by the Source validation webhook.
* Remove the workload reconcilers under
`instrumentor/instrumentationconfig` which were responsible for updating
the instrumentationConfig if the annotation changes.
* Add Source controller which updates InstrumentationConfigs if sources
contain a reported name. This can be unified with the
`startlangdetection` controller which already has a reconciler for
Sources but is currently left out of this PR to keep this one dedicated
to deprecating the annotation.
* Pods mutating webhook is updated to grab the service name based on the
Sources.
* Added a migration to update the Source reported name field from the
deprecated annotation if one exits and the reported name is not
populated in the Source.
* Update the `sources` and `workload-lifecycle` e2e tests. These tests
made use of the annotation to differ between the instrumented service
during different phases of the tests. Both are updated to use the
Sources to achieve that.

---------

Co-authored-by: Mike Dame <mike@odigos.io>