[GEN-2297]: fix patch permission for sources in UI (#2342)
This pull request includes changes to the permissions for managing
sources in the Odigos UI. The main updates involve adding the 'patch'
verb to the list of allowed actions for the `sources` resource.

Changes to permissions:

*
[`cli/cmd/resources/README.md`](diffhunk://#diff-7c3af7b93ef4c36c2e6a5b2866deb5b05c5cb4261ba09ff786c36a0385405bdcL37-R37):
Updated the permissions table to include the 'patch' verb for the
`sources` resource, which is necessary to update the service name.
*
[`cli/cmd/resources/ui.go`](diffhunk://#diff-c286e10d34710a80a59127b2b7951e8a33d9b9554e47d2f2b827fd690f2e53abL311-R315):
Modified the `NewUIClusterRole` function to include the 'patch' verb for
the `sources` resource, ensuring the ability to update the service name.
*
[`helm/odigos/templates/ui/clusterrole.yaml`](diffhunk://#diff-c4bc8c6f6754a6b1d2d13756abd015adc7724832271e98a5b97174a54ee73a90R64):
Added the 'patch' verb to the list of allowed actions for the `sources`
resource in the Helm template, conditional on the `uiMode` not being
"readonly".