List pods to find those relevant for runtime detection instead of caching the workload objects (#2325)
This PR refactors the runtime detection controllers to avoid `Get` or
`List` operation on workload objects (Deployments, StatefulSets and
DaemonSets). As a result:
* The Odiglet permissions are narrowed.
* The cache won't contain workload objects which results in a
significant memory consumption improvement (especially in large
clusters).

This is done by modifying the `insttrumentationconfig` reconciler to
list the Pods associated to the reconciled instrumentationConfig.
The Odiglet cache is already configured to only include Pods whithin the
same node, hence this list operation will only return pods in the same
node and in the same ns of the instrumentationConfig.

---------

Co-authored-by: Amir Blum <amirgiraffe@gmail.com>