Updates for OpenShift compatibility (#2348)
OpenShift seems to enable
[OwnerReferencesPermissionEnforcement](https://kubernetes.io/docs/reference/access-authn-authz/admission-controllers/#ownerreferencespermissionenforcement)
by default, which gives errors like this when setting OwnerReferences:

```
2025-01-29T19:42:38Z	ERROR	Reconciler error	{"controller": "startlangdetection-source", "controllerGroup": "odigos.io", "controllerKind": "Source", "Source": {"name":"source-xsbv2","namespace":"test-project"}, "namespace": "test-project", "name": "source-xsbv2", "reconcileID": "c36405f6-40bb-4eb0-9d5d-fe4bc7802994", "error": "instrumentationconfigs.odigos.io \"deployment-inventory\" is forbidden: cannot set blockOwnerDeletion if an ownerReference refers to a resource you can't set finalizers on: , <nil>"}
```

This basically means wherever we update an ownerreference with
`BlockOwnerDeletion: true` (which is set by the controller-runtime
helpers), we also need permission to update the finalizers for whatever
we're making the owner.

I found this in these places:

* When the autoscaler creates a collector configmap, the owner of the
ConfigMap is the CollectorsGroups. So we need to be able to update
`collectorsgroups/finalizers`
* When the Instrumentor creates an InstrumentationConfig, the owner of
the InstrumentationConfig is the workload. So we need finalizers for
deployments,statefulsets,and daemonsets
* When the odiglet creates an InstrumentationInstance, the owner of the
InstrumentationInstance is also the workload. Same as above

This also adds the SELinux volume mount from
https://github.com/odigos-io/odigos/pull/1131 to the Odiglet init
container added in https://github.com/odigos-io/odigos/pull/1355, along
with making the init container privileged and adding logging for when
the SELinux commands aren't found (this was failing silently)

This also updates the `--openshift` flag to automatically use the
`*-ubi9` images. If no `--image-prefix` is set, the images will be
pulled from `registry.connect.redhat.com/odigos`.