fix: recreate autoscaler resources if deleted or modified (#2409)
If for any reason, the data-collection daemonset, gateway deployment, or
one of the config maps is deleted, then it's not re-created at the
moment.

This PR makes it so they are.

We had a bug where the collectors group would get deleted, and it also
deleted the daemonset. then if the collectors group is re-created
immediately, the daemonset would stay deleted and odigos would not work